Base commit.
This version has a working version of:
 - account signup/creation
it supports a username, email and password. it checks if the username and email exist.

- basic login
it confirms if a password matches a pwd hash and if so: it gives the user a session token. login currently only supports email

- logging out
a username and session token are given to the api, if correct: it deletes the session token from the database thus invalidating that session or "logging out".