uploading and viewing media now works!
You can now view and upload media if you are a registered user.
Go to /media/get/{contentid} to get media.
When you upload media you get given a content id to view it. auth is not needed to view media.
There are currently no media filesize limits. but the types of files allowed to be uploaded IS LIMITED.